分页加载

+++ app/src/main/java/com/example/wschat/db/MessageDao.kt
@@ -14,6 +14,6 @@ interface MessageDao {
     fun insertMessage(vararg messageItem: MessageItem)
 
     @Query("DELETE FROM message where id=(:id)")
-    fun deleteMessage(id: Int)
+    fun deleteMessage(id: Long)
 
 }
